Ya estoy muerto por dentro

+++ Tarea1/pasajero.h
@@ -7,11 +7,11 @@ typedef struct {
     int numeroPasajero;
     int numFila;
     int tipoPasajero; // 0 para economy, 1 para ejecutivo. 2 para internacional
-    double tiempoAbordaje;
+    long long tiempoAbordaje;
     char filaPerteneciente[15]; // "Economy", "Ejecutivo" o "Internacional"
 } Pasajero;
 
-void pasajero_init(Pasajero* this, int numeroPasajero, int numFila, int tipoPasajero, double tiempoAbordaje);
+void pasajero_init(Pasajero* this, int numeroPasajero, int numFila, int tipoPasajero, long long tiempoAbordaje);
 void pasajero_print(const Pasajero* this);
 
 #endif /* PASAJERO_H */
